+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Breakpoint Sync
 
+- Support synchronisation of enabled/disabled state of breakpoints
+
 ## v0.0.1 - 2020-06-29
 
 - Initial release